+++ src/main/resources/application-prod.yml
@@ -7,6 +7,35 @@ spring:
     jdbc:
       initialize-schema: always
 
+  # 세션 설정 (Redis 기반)
+  session:
+    store-type: redis
+    timeout: 30m  # 30분 세션 만료
+
+  # 이메일 설정
+  mail:
+    host: smtp.gmail.com
+    port: 587
+    username: ${MAIL_USERNAME}
+    password: ${MAIL_PASSWORD}
+    properties:
+      mail:
+        smtp:
+          auth: true
+          starttls:
+            enable: true
+            required: true
+        mime:
+          charset: UTF-8
+    default-encoding: UTF-8
+
+  # 스케줄링 활성화
+  task:
+    scheduling:
+      enabled: true
+      pool:
+        size: 2
+
   # 데이터베이스 연결 설정
   datasource: # 여러 DBaaS에서 가져오기
     url: ${DB_URL} # 데이터베이스 URL
@@ -128,3 +157,20 @@ cloud:
       bucket: ${S3_BUCKET_NAME}
     cloudfront:
       domain: ${CLOUD_FRONT_DOMAIN_NAME}
+
+# 관리자 관련 설정
+admin:
+  # 슈퍼관리자 초기 계정 설정
+  super:
+    email: ${ADMIN_SUPER_EMAIL}
+    name: ${ADMIN_SUPER_NAME}
+    password: ${ADMIN_SUPER_PASSWORD}
+
+  # 초대 관련 설정
+  invitation:
+    expiration-hours: ${ADMIN_INVITATION_EXPIRATION_HOURS:1}
+
+# 애플리케이션 설정
+app:
+  admin:
+    base-url: ${APP_BASE_URL}  # 이메일 링크용 (운영 도메인)
